Move the schemas
