+++ laserstudio/config_generator/__init__.py
@@ -1,5 +1,6 @@
-from .config_generator import main
+from .config_generator import main as main_cli
+from .config_generator_wizard import main as main_gui
 from .config_generator import ConfigGenerator
-from .config_generator_ui import ConfigGeneratorWizard
+from .config_generator_wizard import ConfigGeneratorWizard
 
-__all__ = ["main", "ConfigGenerator", "ConfigGeneratorWizard"]
+__all__ = ["main_cli", "main_gui", "ConfigGenerator", "ConfigGeneratorWizard"]
