+++ laserstudio/config_schema/camera.schema.json
