+++ laserstudio/config_schema/instrument.schema.json
