+++ laserstudio/config_schema/laser.schema.json
