+++ laserstudio/config_schema/probe.schema.json
