+++ laserstudio/config_schema/serial.schema.json
