+++ laserstudio/config_schema/stage.schema.json
